requirements.txt
flask==2.2.2
numpy>=1.18.5
scipy>=1.4.1
pandas>=1.1.4
scikit-learn==1.0.2
multiprocess
# Base ------------------------------------------------------------------------
gitpython
ipython  # interactive notebook
matplotlib>=3.2.2
opencv-python>=4.1.1
Pillow>=7.1.2
psutil  # system resources
PyYAML==6.0
requests>=2.23.0

thop>=0.1.1  # FLOPs computation
torch==1.7.0  # see https://pytorch.org/get-started/locally (recommended)
torchvision==0.8.1
tqdm>=4.64.0
# protobuf==3.20.1  # https://github.com/ultralytics/yolov5/issues/8012

# Logging ---------------------------------------------------------------------
tensorboard>=2.4.1
# clearml>=1.2.0
# comet

# Plotting --------------------------------------------------------------------
seaborn>=0.11.0
